==> Step_5_Check_labeled_pdbs.ipynb <==
import pandas as pd
import numpy as np
import os
from tqdm import tqdm
# from dgl import save_graphs, load_graphs
pd.set_option('display.max_rows', 1000+1)
pd.set_option('display.max_columns', 1000+1)

pdb_dir = 'dataset/Godess_final_data/Godess_data_rev/'
pdb_files_list = os.listdir(pdb_dir)

atom_name_list = []
shift_list = []
wierd_list_h2 = []
wierd_list_h2_ci = []
wierd_list_h2_ci = []

ac_list = []

for i in tqdm(range(len(pdb_files_list))):
    
    pdb_name = pdb_files_list[i]
    
    pdb_name_path = os.path.join(pdb_dir, pdb_name)
    
    pdb_f = pd.read_csv(pdb_name_path)
    
    current_atom_names = list(pdb_f['New_Atom_name'].values)
    atom_name_list.extend(current_atom_names)
    
    current_shift_names = list(pdb_f['main_ring_shift'].values)
    shift_list.extend(current_shift_names)
    
    current_ac_atoms = list(pdb_f.loc[pdb_f['Ac_component'] == 1, :]['Atom_name'].values)
    ac_list.extend(current_ac_atoms)
    
    if ('H2' in current_ac_atoms):
        print(pdb_name, i, 'H2 contained in monosaccharide list shoud double check this')
    
    if 'C11' in pdb_f['New_Atom_name'].values:
        
        if pdb_f.loc[pdb_f['New_Atom_name'] == 'C11', :]['main_ring_shift'].values[0] != -1:
#             print(i, pdb_name, len(pdb_f))
            pdb_f_w = pdb_f
            
    if 'glcn' in pdb_f['reformulated_standard_mono'].values:
        wierd_list_h2.extend(list(pdb_f.loc[(pdb_f['reformulated_standard_mono'] == 'glcn') & 
          (pdb_f['Atom_name'] == 'H2'), :]['shift'].values))
        wierd_list_h2_ci.extend(list(pdb_f.loc[(pdb_f['reformulated_standard_mono'] == 'glcn') & 
          (pdb_f['Atom_name'] == 'H2'), :]['trust'].values))

a = pd.read_csv('dataset/Godess_final_data/Godess_data_rev/labeled_all_889.pdb.csv')
a.loc[a['Ac_component'] == 1]

np.unique(ac_list)

glcn_h2 = pd.DataFrame([wierd_list_h2_ci, wierd_list_h2]).T
glcn_h2.columns = ['Trust', 'shift']
glcn_h2.to_csv('glcn_h2.csv', index = False)

glcn_h2.loc[glcn_h2['Trust'] < 50]['Trust'].max()

glcn_h2.loc[glcn_h2['Trust'] > 50]

# pdb_f

df = pd.DataFrame([atom_name_list, shift_list]).T
df = df.drop_duplicates()

df = df.loc[df[1] != -1.0]
df.index = range(len(df))

all_list = np.unique(df[0].values)
all_list

np.unique(df[0].values)

c_list = [i for i in all_list if 'C' in i]
np.array(c_list)

h_list = list(set(all_list) - set(c_list) )
np.array(sorted(h_list))

all_list[np.where('C' in all_list)[0]]
